+++ src/components/swap/AssetSelector.tsx
@@ -22,7 +22,7 @@ import { modalStore } from "../../store/modalStore";
 import { ChainsTooltip } from "./ChainsTooltip";
 import { AvailableChainsSidebar } from "./AvailableChainsSidebar";
 import { AnimatePresence, motion } from "framer-motion";
-import { formatAmount } from "../../utils/utils";
+import { formatBalance } from "../../utils/utils";
 import { useEVMWallet } from "../../hooks/useEVMWallet";
 import { useBitcoinWallet } from "@gardenfi/wallet-connectors";
 import { useStarknetWallet } from "../../hooks/useStarknetWallet";
@@ -161,8 +161,7 @@ export const AssetSelector: FC<props> = ({ onClose }) => {
                     .toNumber()
                 : balance?.toNumber();
           const fiatBalance =
-            formattedBalance &&
-            (Number(formattedBalance) * Number(fiatRate)).toFixed(5);
+            formattedBalance && (formattedBalance * fiatRate).toFixed(5);
 
           return {
             asset,
@@ -430,8 +429,8 @@ export const AssetSelector: FC<props> = ({ onClose }) => {
                               weight="regular"
                               className={`!text-mid-grey`}
                             >
-                              {formatAmount(
-                                Number(formattedBalance),
+                              {formatBalance(
+                                formattedBalance,
                                 0,
                                 Math.min(asset.decimals, BTC.decimals)
                               )}