+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixed
 
+- Updated output error handling and fixed balance formatting.
 - Reduced unnecessary GET requests triggered by reactive useEffect dependencies in useOrderStatus hook.
 
 ## [2.0.1]